review comments

+++ internal/config/config_util.go
@@ -68,7 +68,7 @@ func OverrideWithKernelListCacheTtlFlag(c cliContext, mountConfig *MountConfig, 
 
 // OverrideWithMaxRetryAttemptFlag overwrites the max-retry-attempts config with
 // the max-retry-attempts flag value if the flag is set.
-func OverrideWithMaxRetryAttemptFlag(c cliContext, mountConfig *MountConfig, retries int) {
+func OverrideWithMaxRetryAttemptFlag(c cliContext, mountConfig *MountConfig, retries int64) {
 	if c.IsSet(MaxRetryAttempts) {
 		mountConfig.GCSRetries.MaxRetryAttempts = retries
 	}
